refactor package for spring extension.
